tests/requirements-test.txt
backoff==2.2.1
mock
pytest==7.3.1
google-cloud-storage